+++ src/data/v2/bars.rs
@@ -49,7 +49,7 @@ pub enum Adjustment {
 }
 
 
-/// A GET request to be issued to the /v2/stocks/<symbol>/bars endpoint.
+/// A GET request to be issued to the /v2/stocks/{symbol}/bars endpoint.
 #[derive(Clone, Debug, Eq, PartialEq, Serialize)]
 pub struct BarsReq {
   /// The symbol for which to retrieve market data.
@@ -127,7 +127,7 @@ impl BarsReqInit {
 }
 
 
-/// A market data bar as returned by the /v2/stocks/<symbol>/bars endpoint.
+/// A market data bar as returned by the /v2/stocks/{symbol}/bars endpoint.
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
 #[non_exhaustive]
 pub struct Bar {
@@ -167,7 +167,7 @@ pub struct Bars {
 
 
 Endpoint! {
-  /// The representation of a GET request to the /v2/stocks/<symbol>/bars endpoint.
+  /// The representation of a GET request to the /v2/stocks/{symbol}/bars endpoint.
   pub Get(BarsReq),
   Ok => Bars, [
     /// The market data was retrieved successfully.