+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Added support for historic trade retrieval via `data::v2::trades`
 - Adjusted `data::v2::last_quote` module to work with multiple symbols
+  and renamed it to `last_quotes`
 
 
 0.25.1